Major rework of sequencing system.
The Sequencer class has been significantly simplified by pushing most of the traversal
logic to the syntax tree nodes. The nodes are in a better position to traverse themselves
than the old monolithic Sequencer.

The SequencerState class has been replaced with a Context class. SequencerState
used to build up a parent/child relationship matching the tree structure of the
nodes being traversed, which was overkill. Now the Context uses Hashes to separately
keep track of any information a node needs to store about the traversal, and a
single Sequencer will have a single Context instead of a hierarchy of SequencerStates.