Added support for parallel sequences (counterpoint) with the == operator
Deleted events.rb and moved NoteEvent to values.rb, since this is really
an evaluated value (specifically, a compound value that contains a set of simpler values)

This checkin broke the cosy max object, but that's ok because I'm the only one
who can use it right now. The Midi and MidiFile renderers should still work fine.